requirements.txt
telethon>=1.34.0
PyYAML>=6.0
aiofiles>=23.0